fix(cssc): 31250186 address all comments in public repo PR (#43)
This change addresses all the comments in the [public repo
PR](https://github.com/Azure/azure-cli-extensions/pull/8530).
Final change for public preview.
Test passes, lint and style are successful.

+++ src/acrcssc/azext_acrcssc/helper/_utility.py
@@ -7,10 +7,14 @@ import re
 from knack.log import get_logger
 from datetime import (datetime, timezone)
 import shutil
-from azure.cli.core.azclierror import InvalidArgumentValueError, ResourceNotFoundError
+from azure.cli.core.azclierror import AzCLIError, InvalidArgumentValueError, ResourceNotFoundError
 from azure.mgmt.core.tools import parse_resource_id
-from ._constants import ERROR_MESSAGE_INVALID_TIMESPAN_VALUE, TMP_DRY_RUN_FILE_NAME
-from ._constants import RESOURCE_GROUP
+from ._constants import (
+    CONTINUOUSPATCH_SCHEDULE_MIN_DAYS,
+    CONTINUOUSPATCH_SCHEDULE_MAX_DAYS,
+    ERROR_MESSAGE_INVALID_TIMESPAN_VALUE,
+    RESOURCE_GROUP,
+    TMP_DRY_RUN_FILE_NAME)
 from .._client_factory import cf_acr_tasks
 
 logger = get_logger(__name__)
@@ -34,7 +38,7 @@ def convert_timespan_to_cron(schedule, date_time=None):
     cron_expression = ""
 
     if unit == 'd':  # day of the month
-        if value < 1 or value > 30:
+        if value < CONTINUOUSPATCH_SCHEDULE_MIN_DAYS or value > CONTINUOUSPATCH_SCHEDULE_MAX_DAYS:
             raise InvalidArgumentValueError(error_msg=ERROR_MESSAGE_INVALID_TIMESPAN_VALUE)
         cron_expression = f'{cron_minute} {cron_hour} */{value} * *'
 
@@ -60,20 +64,24 @@ def convert_cron_to_schedule(cron_expression, just_days=False):
 
 
 def create_temporary_dry_run_file(file_location, tmp_folder):
-    templates_path = os.path.dirname(
-        os.path.join(
-            os.path.dirname(
-                os.path.abspath(__file__)),
-            "../templates/"))
+    templates_path = os.path.abspath(os.path.join(
+        os.path.dirname(__file__),
+        "../templates"))
     logger.debug(f"templates_path:  {templates_path}")
 
-    os.makedirs(tmp_folder, exist_ok=True)
-    file_template_copy = templates_path + "/" + TMP_DRY_RUN_FILE_NAME
+    try:
+        os.makedirs(tmp_folder, exist_ok=True)
+        file_template_copy = os.path.join(templates_path, TMP_DRY_RUN_FILE_NAME)
 
-    shutil.copy2(file_template_copy, tmp_folder)
-    shutil.copy2(file_location, tmp_folder)
-    folder_contents = os.listdir(tmp_folder)
-    logger.debug(f"Copied dry run file {folder_contents}")
+        if not os.path.exists(file_template_copy):
+            raise AzCLIError(f"Template file {file_template_copy} does not exist")
+
+        shutil.copy2(file_template_copy, tmp_folder)
+        shutil.copy2(file_location, tmp_folder)
+        folder_contents = os.listdir(tmp_folder)
+        logger.debug(f"Copied dry run file {folder_contents}")
+    except Exception as exception:
+        raise AzCLIError(f"Failed to copy dry run file: {exception}")
 
 
 def delete_temporary_dry_run_file(tmp_folder):
